+++ jeap-spring-boot-tx/src/main/java/ch/admin/bit/jeap/db/tx/config/JeapTxTransactionAutoConfig.java
@@ -6,15 +6,24 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.support.BeanDefinitionRegistryPostProcessor;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.context.properties.bind.Bindable;
+import org.springframework.boot.context.properties.bind.Binder;
 import org.springframework.boot.transaction.autoconfigure.TransactionAutoConfiguration;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Role;
+import org.springframework.core.env.Environment;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.TransactionManager;
 import org.springframework.transaction.interceptor.TransactionAttributeSource;
 import org.springframework.transaction.interceptor.TransactionInterceptor;
 
+import static org.springframework.beans.factory.config.BeanDefinition.ROLE_INFRASTRUCTURE;
+
 @AutoConfiguration(after = TransactionAutoConfiguration.class)
 public class JeapTxTransactionAutoConfig {
+
+    private JeapTxTransactionAutoConfig() {
+    }
 
     /**
      * Must be static! Otherwise, the whole autoconfiguration class is eagerly instantiated when running bean post
@@ -35,14 +44,19 @@ public class JeapTxTransactionAutoConfig {
     }
 
     @Bean
+    @Role(ROLE_INFRASTRUCTURE)
     @ConditionalOnBean({PlatformTransactionManager.class, TransactionAttributeSource.class})
-    AwsJdbcFailoverRetryAspect awsJdbcFailoverRetryAspect(
+    static AwsJdbcFailoverRetryAdvisor awsJdbcFailoverRetryAdvisor(
             ListableBeanFactory beanFactory,
             TransactionAttributeSource transactionAttributeSource,
+            Environment environment,
             @Qualifier("transactionInterceptor") ObjectProvider<TransactionInterceptor> transactionInterceptorProvider) {
         TransactionInterceptor transactionInterceptor = transactionInterceptorProvider.getIfAvailable();
         TransactionManager defaultTransactionManager =
                 transactionInterceptor == null ? null : transactionInterceptor.getTransactionManager();
-        return new AwsJdbcFailoverRetryAspect(beanFactory, transactionAttributeSource, defaultTransactionManager);
+        AwsJdbcFailoverRetryProperties properties = Binder.get(environment).bindOrCreate(
+                "jeap.datasource.aws.failover-retry", Bindable.of(AwsJdbcFailoverRetryProperties.class));
+        return new AwsJdbcFailoverRetryAdvisor(
+                beanFactory, transactionAttributeSource, defaultTransactionManager, properties);
     }
 }